requirements.txt
psutil>=5.9.0
scapy>=2.5.0
python-nmap>=0.7.1
fpdf2>=2.7.0
python-docx>=1.1.0
pyyaml>=6.0
customtkinter>=5.2
fastapi>=0.115
uvicorn>=0.30
sqlalchemy>=2.0
aiohttp>=3.10
reportlab>=4.2
